refactor: more consistent naming
