--- Scenes/UI/GameUI/SensorDisplay.tscn
@@ -1,63 +1,0 @@
-[gd_scene load_steps=7 format=3 uid="uid://dika23l4ux5px"]
-
-[ext_resource type="PackedScene" uid="uid://k1xfbpv7ywee" path="res://Scenes/UI/StationMenu/IconButton/icon_button.tscn" id="1_f5gu3"]
-[ext_resource type="Script" path="res://Scenes/UI/GameUI/SensorDisplay.cs" id="1_q2mlm"]
-[ext_resource type="Texture2D" uid="uid://7wdkkbckbsb3" path="res://Assets/Icons/exclamation-solid.svg" id="2_bl7nt"]
-[ext_resource type="Texture2D" uid="uid://bl2vqnpjvfs2v" path="res://Assets/Icons/satellite-solid.svg" id="3_gbd2d"]
-[ext_resource type="Texture2D" uid="uid://dfljdr01r3278" path="res://Assets/Icons/location-arrow-solid.svg" id="4_boqwr"]
-[ext_resource type="Texture2D" uid="uid://cleipdwqse37" path="res://Assets/Icons/infinity-solid.svg" id="5_64rif"]
-
-[node name="SensorDisplay" type="PanelContainer"]
-anchors_preset = 2
-anchor_top = 1.0
-anchor_bottom = 1.0
-offset_top = -244.0
-offset_right = 172.0
-grow_vertical = 0
-size_flags_horizontal = 0
-size_flags_vertical = 8
-script = ExtResource("1_q2mlm")
-
-[node name="VBoxContainer" type="VBoxContainer" parent="."]
-layout_mode = 2
-
-[node name="HBoxContainer" type="HBoxContainer" parent="VBoxContainer"]
-layout_mode = 2
-alignment = 1
-
-[node name="Important" parent="VBoxContainer/HBoxContainer" instance=ExtResource("1_f5gu3")]
-unique_name_in_owner = true
-custom_minimum_size = Vector2(20, 20)
-layout_mode = 2
-_iconTexture = ExtResource("2_bl7nt")
-_iconColor = Color(0.5, 0.908333, 1, 1)
-
-[node name="Stations" parent="VBoxContainer/HBoxContainer" instance=ExtResource("1_f5gu3")]
-unique_name_in_owner = true
-layout_mode = 2
-_iconTexture = ExtResource("3_gbd2d")
-_iconColor = Color(0.501961, 0.909804, 1, 1)
-
-[node name="Navigation" parent="VBoxContainer/HBoxContainer" instance=ExtResource("1_f5gu3")]
-unique_name_in_owner = true
-layout_mode = 2
-_iconTexture = ExtResource("4_boqwr")
-_iconColor = Color(0.501961, 0.909804, 1, 1)
-
-[node name="All" parent="VBoxContainer/HBoxContainer" instance=ExtResource("1_f5gu3")]
-unique_name_in_owner = true
-layout_mode = 2
-_iconTexture = ExtResource("5_64rif")
-_iconColor = Color(0.501961, 0.909804, 1, 1)
-
-[node name="ScrollContainer" type="ScrollContainer" parent="VBoxContainer"]
-custom_minimum_size = Vector2(160, 200)
-layout_mode = 2
-horizontal_scroll_mode = 0
-
-[node name="ObjectList" type="VBoxContainer" parent="VBoxContainer/ScrollContainer"]
-unique_name_in_owner = true
-custom_minimum_size = Vector2(160, 200)
-layout_mode = 2
-size_flags_horizontal = 3
-size_flags_vertical = 3
